requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==2.2.19
django-environ==0.4.5
django-redis==4.12.1
djangorestframework==3.12.4
idna==2.10
psycopg2==2.8.6
pytz==2021.1
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
